requirements.txt
Flask==3.0.0
gunicorn
psycopg2-binary
python-dotenv
Flask-WTF
cloudinary